Setup some settings for to use during development, tested the __init__ stuff to make sure it would work.
For some reason, device/__init__.py is called twice. This might be a problem. I don't know
if it is being called twice from one invocation, or if it is being run twice. We may need
to have a static variable (?) that allows to track if this code has been run yet.


git-svn-id: svn://dunari.cis.vtc.edu/POWER@94 f2234473-aea0-4c34-a049-d44cbc1b6343